refactor(eam): Move EAM to services
