+++ src/config/dependencyConfigSTFC.ts
@@ -25,7 +25,7 @@ import {
   createSkipListeningHandler,
   createSkipPostingHandler,
 } from '../eventHandlers/messageBroker';
-import { SkipAssetRegistrar } from '../utils/EAM_service';
+import { SkipAssetRegistrar } from '../services/eam';
 import { configureSTFCEnvironment } from './stfc/configureSTFCEnvironment';
 import { Tokens } from './Tokens';
 import { mapClass, mapValue } from './utils';
